Separated timeseries module into files
